% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_function.R
\name{example_function}
\alias{example_function}
\title{Example Function}
\usage{
example_function(data)
}
\arguments{
\item{data}{Input data, unprocessed.}
}
\description{
Calculates the product of the numbers of rows and columns of a matrix.
}
\author{
Sebastian Schweer
}
