% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_script.R
\name{main}
\alias{main}
\title{Main Function Bundling it all together}
\usage{
main(data)
}
\arguments{
\item{data}{Input data, unprocessed.}
}
\description{
Applies the main functions to DATA.
}
\author{
Sebastian Schweer
}
